requirements.txt
omegaconf==2.1.1
numpy==1.24.1
pillow==9.3.0
tqdm==4.66.1
scipy==1.11.4
matplotlib==3.8.2
ninja==1.10.2